fix: subclass MDAWidget instead
